Pull out role db requests into role model
